rename: Sidebar -> DetailSideTable 이름 변경
