+++ src/App.tsx
@@ -3,8 +3,8 @@ import ChickenTable from './components/ChickenTable/ChickenTable';
 import ProductsFilter from './components/Products/ProductsFilter';
 import { ChickenData } from './types/ChickenData';
 import DemoPage from "./components/DemoPage";
-import Sidebar from "./components/Sidebar/Sidebar";
 import './App.css';
+import DetailSideTable from "./components/Details/DetailSideTable";
 
 const App: React.FC = () => {
   const [filteredData, setFilteredData] = useState<ChickenData[]>([]);
@@ -14,7 +14,7 @@ const App: React.FC = () => {
       <DemoPage />
       <ProductsFilter onFilter={setFilteredData} />
       <ChickenTable filteredData={filteredData} />
-      <Sidebar />
+      <DetailSideTable />
     </div>
   );
 }
