+++ src/components/Details/DetailSideTable.css.ts
